requirements.txt
# Testing & Code Quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
codecov>=2.1.13

# Add your project dependencies below
openai>=1.0.0
python-dotenv>=1.0.0
click>=8.1.0
requests>=2.31.0
